Moved: `modifier.py` to `src/pytorch/sparsification`
