+++ src/sparseml/pytorch/sparsification/__init__.py
@@ -25,3 +25,4 @@ from .info import *
 from .modifier_thinning import *
 from .pruning import *
 from .quantization.modifier_quantization import *
+from .modifier import *
