+++ src/sparseml/pytorch/sparsification/pruning/modifier_pruning_acdc.py
@@ -18,7 +18,7 @@ from typing import List, Union
 from torch.nn import Module, Parameter
 from torch.optim.optimizer import Optimizer
 
-from sparseml.pytorch.optim.modifier import ModifierProp, PyTorchModifierYAML
+from sparseml.pytorch.sparsification.modifier import ModifierProp, PyTorchModifierYAML
 from sparseml.pytorch.sparsification.pruning.mask_creator import (
     PruningMaskCreator,
     get_mask_creator_default,
