+++ src/sparseml/pytorch/sparsification/pruning/modifier_pruning_structured.py
@@ -23,7 +23,7 @@ import torch
 from torch import Tensor
 from torch.nn import Parameter
 
-from sparseml.pytorch.optim.modifier import ModifierProp, PyTorchModifierYAML
+from sparseml.pytorch.sparsification.modifier import ModifierProp, PyTorchModifierYAML
 from sparseml.pytorch.sparsification.pruning.mask_creator import (
     GroupedPruningMaskCreator,
     PruningMaskCreator,
